+++ web/static/styles.css
