+++ web/templates/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Calculadora de Vigas</title>
-    <link rel="stylesheet" href="/static/styles.css">
+    <link rel="stylesheet" href="/web/static/styles.css">
 </head>
 <body>
     <div class="container">
@@ -55,6 +55,6 @@
         </div>        
     </div>
 
-    <script src="/static/scripts.js"></script>
+    <script src="/web/static/scripts.js"></script>
 </body>
 </html>
